Moved windows code -> windows_helper, 'search_for_log' -> let folderpath is no longer mut
